SRE/fastapi-server/assets/templates/requirements.txt
# Core FastAPI dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0

# Optional: Add based on your needs
# Database
# sqlalchemy==2.0.36
# asyncpg==0.30.0  # For PostgreSQL
# psycopg2-binary==2.9.10  # Alternative PostgreSQL driver

# Authentication
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4
# python-multipart==0.0.18

# HTTP Client
# httpx==0.28.0
# aiohttp==3.11.7

# Testing
# pytest==8.3.4
# pytest-asyncio==0.24.0
# httpx==0.28.0  # For TestClient

# Development
# black==24.10.0
# ruff==0.8.4
# mypy==1.13.0
